+++ frontend/src/lib/components/ui/gallery/index.ts
@@ -1,6 +1,6 @@
 import type { Loading, MediaDetails, GalleryItem } from "./types.ts";
 
-export { default as Gallery2 } from "./gallery-2.svelte";
-export { default as Media2 } from "./media-2.svelte";
+export { default as Gallery } from "./gallery.svelte";
+export { default as Media } from "./media.svelte";
 
 export type { MediaDetails, GalleryItem, Loading };
