+++ frontend/src/lib/components/ui/gallery/media.svelte
